--- EAD Assinment 6(2)/.settings/org.eclipse.wst.common.component
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="Soap">
-        <wb-resource deploy-path="/" source-path="/WebContent" tag="defaultRootSource"/>
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src"/>
-        <property name="context-root" value="Soap"/>
-        <property name="java-output-path" value="/Soap/build/classes"/>
-    </wb-module>
-</project-modules>
