--- EAD Assinment 6(2)/WebContent/WEB-INF/AreaTriangleService/abhinav/deploy.wsdd
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><!-- Use this file to deploy some handlers/chains and services      --><!-- Two ways to do this:                                           --><!--   java org.apache.axis.client.AdminClient deploy.wsdd          --><!--      after the axis server is running                          --><!-- or                                                             --><!--   java org.apache.axis.utils.Admin client|server deploy.wsdd   --><!--      from the same directory that the Axis engine runs         --><deployment xmlns="http://xml.apache.org/axis/wsdd/" xmlns:java="http://xml.apache.org/axis/wsdd/providers/java">
-
-  <!-- Services from AreaTriangleService WSDL service -->
-
-  <service name="AreaTriangle" provider="java:RPC" style="wrapped" use="literal">
-      <parameter name="wsdlTargetNamespace" value="http://abhinav"/>
-      <parameter name="wsdlServiceElement" value="AreaTriangleService"/>
-      <parameter name="schemaQualified" value="http://abhinav"/>
-      <parameter name="wsdlServicePort" value="AreaTriangle"/>
-      <parameter name="className" value="abhinav.AreaTriangle"/>
-      <parameter name="wsdlPortType" value="AreaTriangle"/>
-      <parameter name="typeMappingVersion" value="1.2"/>
-      <operation xmlns:operNS="http://abhinav" xmlns:retNS="http://abhinav" xmlns:rtns="http://www.w3.org/2001/XMLSchema" name="area" qname="operNS:area" returnQName="retNS:areaReturn" returnType="rtns:double" soapAction="">
-        <parameter xmlns:pns="http://abhinav" xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="pns:sideA" type="tns:double"/>
-        <parameter xmlns:pns="http://abhinav" xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="pns:sideB" type="tns:double"/>
-        <parameter xmlns:pns="http://abhinav" xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="pns:sideC" type="tns:double"/>
-      </operation>
-      <parameter name="allowedMethods" value="area"/>
-
-  </service>
-</deployment>
