--- EAD Assinment 6(2)/WebContent/WEB-INF/AreaTriangleService/abhinav/deploy.wsdd.bak
@@ -1,31 +1,0 @@
-<!-- Use this file to deploy some handlers/chains and services      -->
-<!-- Two ways to do this:                                           -->
-<!--   java org.apache.axis.client.AdminClient deploy.wsdd          -->
-<!--      after the axis server is running                          -->
-<!-- or                                                             -->
-<!--   java org.apache.axis.utils.Admin client|server deploy.wsdd   -->
-<!--      from the same directory that the Axis engine runs         -->
-
-<deployment
-    xmlns="http://xml.apache.org/axis/wsdd/"
-    xmlns:java="http://xml.apache.org/axis/wsdd/providers/java">
-
-  <!-- Services from AreaTriangleService WSDL service -->
-
-  <service name="AreaTriangle" provider="java:RPC" style="wrapped" use="literal">
-      <parameter name="wsdlTargetNamespace" value="http://abhinav"/>
-      <parameter name="wsdlServiceElement" value="AreaTriangleService"/>
-      <parameter name="schemaQualified" value="http://abhinav"/>
-      <parameter name="wsdlServicePort" value="AreaTriangle"/>
-      <parameter name="className" value="abhinav.AreaTriangleSoapBindingImpl"/>
-      <parameter name="wsdlPortType" value="AreaTriangle"/>
-      <parameter name="typeMappingVersion" value="1.2"/>
-      <operation name="area" qname="operNS:area" xmlns:operNS="http://abhinav" returnQName="retNS:areaReturn" xmlns:retNS="http://abhinav" returnType="rtns:double" xmlns:rtns="http://www.w3.org/2001/XMLSchema" soapAction="" >
-        <parameter qname="pns:sideA" xmlns:pns="http://abhinav" type="tns:double" xmlns:tns="http://www.w3.org/2001/XMLSchema"/>
-        <parameter qname="pns:sideB" xmlns:pns="http://abhinav" type="tns:double" xmlns:tns="http://www.w3.org/2001/XMLSchema"/>
-        <parameter qname="pns:sideC" xmlns:pns="http://abhinav" type="tns:double" xmlns:tns="http://www.w3.org/2001/XMLSchema"/>
-      </operation>
-      <parameter name="allowedMethods" value="area"/>
-
-  </service>
-</deployment>
