--- EAD Assinment 6(2)/WebContent/wsdl/AreaTriangle.wsdl
@@ -1,92 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<wsdl:definitions targetNamespace="http://abhinav" xmlns:apachesoap="http://xml.apache.org/xml-soap" xmlns:impl="http://abhinav" xmlns:intf="http://abhinav" xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/" xmlns:wsdlsoap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
-<!--WSDL created by Apache Axis version: 1.4
-Built on Apr 22, 2006 (06:55:48 PDT)-->
- <wsdl:types>
-  <schema elementFormDefault="qualified" targetNamespace="http://abhinav" xmlns="http://www.w3.org/2001/XMLSchema">
-   <element name="area">
-    <complexType>
-     <sequence>
-      <element name="sideA" type="xsd:double"/>
-      <element name="sideB" type="xsd:double"/>
-      <element name="sideC" type="xsd:double"/>
-     </sequence>
-    </complexType>
-   </element>
-   <element name="areaResponse">
-    <complexType>
-     <sequence>
-      <element name="areaReturn" type="xsd:double"/>
-     </sequence>
-    </complexType>
-   </element>
-  </schema>
- </wsdl:types>
-
-   <wsdl:message name="areaResponse">
-
-      <wsdl:part element="impl:areaResponse" name="parameters">
-
-      </wsdl:part>
-
-   </wsdl:message>
-
-   <wsdl:message name="areaRequest">
-
-      <wsdl:part element="impl:area" name="parameters">
-
-      </wsdl:part>
-
-   </wsdl:message>
-
-   <wsdl:portType name="AreaTriangle">
-
-      <wsdl:operation name="area">
-
-         <wsdl:input message="impl:areaRequest" name="areaRequest">
-
-       </wsdl:input>
-
-         <wsdl:output message="impl:areaResponse" name="areaResponse">
-
-       </wsdl:output>
-
-      </wsdl:operation>
-
-   </wsdl:portType>
-
-   <wsdl:binding name="AreaTriangleSoapBinding" type="impl:AreaTriangle">
-
-      <wsdlsoap:binding style="document" transport="http://schemas.xmlsoap.org/soap/http"/>
-
-      <wsdl:operation name="area">
-
-         <wsdlsoap:operation soapAction=""/>
-
-         <wsdl:input name="areaRequest">
-
-            <wsdlsoap:body use="literal"/>
-
-         </wsdl:input>
-
-         <wsdl:output name="areaResponse">
-
-            <wsdlsoap:body use="literal"/>
-
-         </wsdl:output>
-
-      </wsdl:operation>
-
-   </wsdl:binding>
-
-   <wsdl:service name="AreaTriangleService">
-
-      <wsdl:port binding="impl:AreaTriangleSoapBinding" name="AreaTriangle">
-
-         <wsdlsoap:address location="http://localhost:8080/Soap/services/AreaTriangle"/>
-
-      </wsdl:port>
-
-   </wsdl:service>
-
-</wsdl:definitions>
